Implemented Adding Movie Review
- Fixed some syntax errors in movies.sql that were giving errors when
you source.
- Added a form to write review page that adds a movie
information/review into the database once sumbitted.
- Will figure out how to force user to fill out all the form fields
before sumbitting sometime later

NOTE: I was having problems logging in as the admin to access the
'writeReview' page. When I logged in (username: admin & password:
54321), it sent me to the page however there was a blank screen. I
figured this was happing because of the exit statement that runs if the
login cookie is not set (i have commented it out for the time being and
for testing). Therefore, there seems to be a problem with it setting
the login cookie once you login. This may just be a problem that I'm
experiencing, not sure. Will try to look into it more later.